Revert "WT-12938 Skip compact for small files in background compact (#10633)"
This reverts commit f5be21f36eb5a01023461ff7e4d0adc5958cd40c.
