+++ test/suite/compact_util.py
@@ -72,12 +72,6 @@ class compact_util(wttest.WiredTigerTestCase):
     def get_pages_rewritten(self, uri):
         return self.get_stat(stat.dsrc.btree_compact_pages_rewritten, uri)
 
-    def get_bg_compaction_files_skipped(self):
-        stat_cursor = self.session.open_cursor('statistics:', None, None)
-        skipped = stat_cursor[stat.conn.background_compact_skipped][2]
-        stat_cursor.close()
-        return skipped
-
     # Return the size of the given file.
     def get_size(self, uri):
         # To allow this to work on systems without ftruncate,
