+++ test/suite/test_compact06.py
@@ -36,6 +36,12 @@ from wiredtiger import stat
 class test_compact06(compact_util):
     configuration_items = ['exclude=["table:a.wt"]', 'free_space_target=10MB', 'timeout=60']
 
+    def get_bg_compaction_files_skipped(self):
+        stat_cursor = self.session.open_cursor('statistics:', None, None)
+        skipped = stat_cursor[stat.conn.background_compact_skipped][2]
+        stat_cursor.close()
+        return skipped
+
     def test_background_compact_api(self):
         # FIXME-WT-11399
         if self.runningHook('tiered'):
@@ -67,35 +73,38 @@ class test_compact06(compact_util):
                 self.session.compact(None, f'background=true,{item}'),
                 '/Cannot reconfigure background compaction while it\'s already running/')
 
-        # Wait for background compaction to start and skip the HS.
-        while self.get_bg_compaction_files_skipped() == 0:
+        # Wait for background compaction to process the HS.
+        while self.get_bg_compaction_success() == 0:
             time.sleep(1)
 
         # Disable the background compaction server.
         self.turn_off_bg_compact()
 
-        # Background compaction should have skipped the HS file as it is less than 1MB.
-        assert self.get_bg_compaction_files_skipped() == 1
+        # Background compaction should have tried to compact the HS hence skipped no files.
+        assert self.get_bg_compaction_files_skipped() == 0
+        assert self.get_bg_compaction_success() == 1
 
         # Enable background and configure it to run once. Don't use the helper function as the
         # server may go to sleep before we have the time to check it is actually running.
         self.session.compact(None, 'background=true,run_once=true')
 
-        # Wait for background compaction to start and skip the HS file again.
-        while self.get_bg_compaction_files_skipped() == 1:
+        # Wait for background compaction to process the HS table. Even though there is no work to
+        # do, it is considered as a success.
+        while self.get_bg_compaction_success() == 1:
             time.sleep(1)
 
         # Ensure background compaction stops by itself.
         while self.get_bg_compaction_running():
             time.sleep(1)
 
-        # Background compact should only skip the HS file once.
-        assert self.get_bg_compaction_files_skipped() == 2
+        # When running once, background compaction should not skip files.
+        assert self.get_bg_compaction_files_skipped() == 0
+        assert self.get_bg_compaction_success() == 2
 
         # Enable the server again but with default options, the HS should be skipped.
         self.turn_on_bg_compact()
 
-        while self.get_bg_compaction_files_skipped() == 2:
+        while self.get_bg_compaction_files_skipped() != 1:
             time.sleep(1)
 
         self.turn_off_bg_compact()
